backend/requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
fastapi==0.96.0
filelock==3.12.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
pydantic==1.10.8
PyMySQL==1.0.3
python-multipart==0.0.6
requests==2.31.0
schedule==1.2.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.3
urllib3==2.0.2
uvicorn==0.22.0
